# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Backlogg API",
    "description": "Backlogg is a unified catalog API for **movies, series, books and games**,\nseeded from public external APIs (TMDB, Open Library, IGDB) via a nightly sync,\nplus a social layer on top of that catalog.\n\n**Catalog & discovery:** list and detail endpoints per content type, cross-type\n`/search` (with on-demand external fallback), `/genres`, `/trending`, `/people`\nand similar-item recommendations.\n\n**Social layer:** account management with refresh tokens and account recovery\n(`/auth`, `/users`), ratings & reviews, per-user library/backlog, curated\n`/lists`, `/follows`, an activity `/feed`, `/notifications` and personalized\n`/recommendations`.\n\n**Platform:** rate limiting, structured observability, Prometheus `/metrics`\nand response caching.\n\n## Authentication\n\n- **User endpoints** use `Authorization: Bearer <access_token>`. Obtain the\n  token pair from `POST /auth/login`; renew it with `POST /auth/refresh`.\n- **Admin endpoints** (`/admin/*`) require the `X-API-Key` header.\n\nEverything else is public. Direct messaging between users is out of scope.\n",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/movies": {
      "get": {
        "tags": [
          "movies"
        ],
        "summary": "List movies",
        "description": "Paginated list of catalogued movies (no external fallback); genre filter + sort.",
        "operationId": "list_movies_v1_movies_get",
        "parameters": [
          {
            "name": "genre",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by genre slug",
              "title": "Genre"
            },
            "description": "Filter by genre slug"
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/MovieSortEnum",
              "description": "Sort order",
              "default": "rating_desc"
            },
            "description": "Sort order"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Items per page",
              "default": 20,
              "title": "Limit"
            },
            "description": "Items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MovieListOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/movies/{slug}/similar": {
      "get": {
        "tags": [
          "movies"
        ],
        "summary": "Similar movies",
        "description": "Up to 10 similar movies (TMDB). New items are persisted locally.",
        "operationId": "get_similar_movies_v1_movies__slug__similar_get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimilarMoviesOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/movies/{slug}/ratings": {
      "get": {
        "tags": [
          "movies"
        ],
        "summary": "List movie ratings",
        "description": "Public, paginated ratings & reviews for a movie, newest first.",
        "operationId": "list_movie_ratings_v1_movies__slug__ratings_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Items per page",
              "default": 20,
              "title": "Limit"
            },
            "description": "Items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RatingListOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/movies/{slug}/rating": {
      "put": {
        "tags": [
          "movies"
        ],
        "summary": "Rate a movie",
        "description": "Upsert the caller's score/review (full replace); recomputes aggregates. Auth.",
        "operationId": "rate_movie_v1_movies__slug__rating_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RatingIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RatingOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "movies"
        ],
        "summary": "Delete own movie rating",
        "description": "Remove the caller's rating and recompute aggregates. Requires auth.",
        "operationId": "delete_movie_rating_v1_movies__slug__rating_delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/movies/{slug}/library": {
      "put": {
        "tags": [
          "movies"
        ],
        "summary": "Set movie library status",
        "description": "Upsert the caller's backlog status for this movie. Requires auth.",
        "operationId": "set_movie_library_v1_movies__slug__library_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LibraryEntryIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryStatusOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "movies"
        ],
        "summary": "Remove movie from library",
        "description": "Delete the caller's library entry for this movie. Requires auth.",
        "operationId": "delete_movie_library_v1_movies__slug__library_delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/movies/{slug}": {
      "get": {
        "tags": [
          "movies"
        ],
        "summary": "Get movie detail",
        "description": "Full movie detail; on-demand fallback. Auth optional (adds viewer_status).",
        "operationId": "get_movie_v1_movies__slug__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MovieOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GenreOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "slug"
        ],
        "title": "GenreOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LibraryEntryIn": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/LibraryStatus"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "LibraryEntryIn",
        "description": "Request body for PUT /{type}/{slug}/library."
      },
...
      "LibraryStatus": {
        "type": "string",
        "enum": [
          "want",
          "in_progress",
          "completed",
          "dropped"
        ],
        "title": "LibraryStatus",
        "description": "Backlog status values accepted in request bodies and filters."
      },
      "LibraryStatusOut": {
        "properties": {
          "item_type": {
            "type": "string",
            "title": "Item Type"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "item_type",
          "slug",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "LibraryStatusOut",
        "description": "Response of PUT /{type}/{slug}/library \u2014 the caller's own entry.",
        "example": {
          "created_at": "2026-05-25T18:04:11Z",
          "item_type": "MOVIE",
          "slug": "dune-2021",
          "status": "want",
          "updated_at": "2026-05-25T18:04:11Z"
        }
      },
...
      "MovieListItemOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "poster_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Poster Url"
          },
          "release_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Release Date"
          },
          "rating_external": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating External"
          },
          "genres": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Genres"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "slug",
          "poster_url",
          "release_date",
          "rating_external",
          "genres"
        ],
        "title": "MovieListItemOut"
      },
      "MovieListOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MovieListItemOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "limit"
        ],
        "title": "MovieListOut",
        "example": {
          "items": [
            {
              "genres": [
                "science-fiction",
                "adventure"
              ],
              "id": 1,
              "poster_url": "https://image.tmdb.org/t/p/w500/dune.jpg",
              "rating_external": 7.8,
              "release_date": "2021-10-22",
              "slug": "dune-2021",
              "title": "Dune"
            }
          ],
          "limit": 20,
          "page": 1,
          "total": 847
        }
      },
      "MovieOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "original_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Title"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "overview": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overview"
          },
          "release_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Release Date"
          },
          "runtime": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Runtime"
          },
          "original_language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Language"
          },
          "poster_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Poster Url"
          },
          "backdrop_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Backdrop Url"
          },
          "budget": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Budget"
          },
          "revenue": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revenue"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "rating_external": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating External"
          },
          "rating_count_external": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating Count External"
          },
          "rating_internal": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating Internal"
          },
          "rating_count_internal": {
            "type": "integer",
            "title": "Rating Count Internal"
          },
          "genres": {
            "items": {
              "$ref": "#/components/schemas/GenreOut"
            },
            "type": "array",
            "title": "Genres"
          },
          "credits": {
            "items": {
              "$ref": "#/components/schemas/backlogg__shared__schemas__CreditOut"
            },
            "type": "array",
            "title": "Credits",
            "default": []
          },
          "viewer_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Viewer Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "original_title",
          "slug",
          "overview",
          "release_date",
          "runtime",
          "original_language",
          "poster_url",
          "backdrop_url",
          "budget",
          "revenue",
          "status",
          "rating_external",
          "rating_count_external",
          "rating_internal",
          "rating_count_internal",
          "genres"
        ],
        "title": "MovieOut",
        "example": {
          "backdrop_url": "https://image.tmdb.org/t/p/w780/dune-bd.jpg",
          "budget": 165000000,
          "credits": [
            {
              "billing_order": 0,
              "character_name": "Paul Atreides",
              "person_name": "Timoth\u00e9e Chalamet",
              "person_slug": "timothee-chalamet",
              "profile_url": "https://image.tmdb.org/t/p/w185/tc.jpg",
              "role": "cast"
            }
          ],
          "genres": [
            {
              "id": 3,
              "name": "Science Fiction",
              "slug": "science-fiction"
            }
          ],
          "id": 1,
          "original_language": "en",
          "original_title": "Dune",
          "overview": "Paul Atreides unites with the Fremen of Arrakis.",
          "poster_url": "https://image.tmdb.org/t/p/w500/dune.jpg",
          "rating_count_external": 9231,
          "rating_count_internal": 87,
          "rating_external": 7.8,
          "rating_internal": 4.2,
          "release_date": "2021-10-22",
          "revenue": 401800000,
          "runtime": 155,
          "slug": "dune-2021",
          "status": "Released",
          "title": "Dune",
          "viewer_status": "want"
        }
      },
      "MovieSortEnum": {
        "type": "string",
        "enum": [
          "rating_desc",
          "rating_asc",
          "date_desc",
          "date_asc",
          "title_asc"
        ],
        "title": "MovieSortEnum"
      },
...
      "RatingAuthorOut": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          }
        },
        "type": "object",
        "required": [
          "username",
          "display_name",
          "avatar_url"
        ],
        "title": "RatingAuthorOut"
      },
      "RatingIn": {
        "properties": {
          "score": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 5.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "review_text": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 10000
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Text"
          }
        },
        "type": "object",
        "title": "RatingIn",
        "description": "Request body for PUT /{type}/{slug}/rating \u2014 full replace (upsert)."
      },
      "RatingListOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RatingOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "limit"
        ],
        "title": "RatingListOut"
      },
      "RatingOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "user": {
            "$ref": "#/components/schemas/RatingAuthorOut"
          },
          "score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "review_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Text"
          },
          "like_count": {
            "type": "integer",
            "title": "Like Count"
          },
          "liked_by_viewer": {
            "type": "boolean",
            "title": "Liked By Viewer"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user",
          "score",
          "review_text",
          "like_count",
          "liked_by_viewer",
          "created_at",
          "updated_at"
        ],
        "title": "RatingOut"
      },
...
      "SimilarMovieOut": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "poster_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Poster Url"
          },
          "release_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Release Date"
          },
          "rating_external": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating External"
          }
        },
        "type": "object",
        "required": [
          "title",
          "slug",
          "poster_url",
          "release_date",
          "rating_external"
        ],
        "title": "SimilarMovieOut"
      },
      "SimilarMoviesOut": {
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/SimilarMovieOut"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "type": "object",
        "required": [
          "results"
        ],
        "title": "SimilarMoviesOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}